Refactor: CounselingCenter, Comments Screen & Update: refactoring screen style.js file
