+++ screens/CounselingCenter/CategoryPosts/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  postsWrapper: {
+    width: "100%",
+    justifyContent: "center",
+    alignItems: "center",
+    marginTop: 0,
+    marginRight: "auto",
+    marginBottom: 0,
+    marginLeft: "auto"
+  },
+  emptyContainer: {
+    marginTop: "15%"
+  }
+});
+
+export default styles;
